Title: 사각형 채우기 3, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 260226/사각형 채우기 3/rectangle-fill-3.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 
 int N;
-vector<int> dp;
+vector<long long> dp;
 
 void init()
 {
